requirements.txt
Flask
numpy
pandas
tensorflow
scikit-learn
google-cloud-storage
google-cloud-secret-manager